+++ foxes/algorithms/__init__.py
@@ -4,7 +4,7 @@ Algorithms define the main calculation routines.
 
 from .downwind.downwind import Downwind
 from .iterative.iterative import Iterative
-from .sequential.sequential import Sequential
+from .sequential import Sequential
 
 from . import downwind
 from . import iterative