+++ foxes/output/seq_plugins/__init__.py
@@ -1,0 +1,2 @@
+from .seq_flow_ani_plugin import SeqFlowAnimationPlugin
+from .seq_wake_debug_plugin import SeqWakeDebugPlugin